+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/onboarding/MicOnboardingScreen.java
@@ -19,7 +19,7 @@ public class MicOnboardingScreen extends DeviceOnboardingScreen {
 
     @Override
     public AudioDeviceList createAudioDeviceList(int width, int height, int top) {
-        return new MicrophoneAudioDeviceList(width, height, top);
+        return new MicrophoneAudioDeviceList(this, width, height, top);
     }
 
     @Override