+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/widgets/MicTestButton.java
@@ -29,12 +29,15 @@ public class MicTestButton extends ToggleImageButton implements ImageButton.Tool
     private boolean micActive;
     @Nullable
     private VoiceThread voiceThread;
+    @Nullable
     private final MicListener micListener;
+    private final boolean raw;
     @Nullable
     private final ClientVoicechat client;
 
-    public MicTestButton(int xIn, int yIn, MicListener micListener) {
+    public MicTestButton(int xIn, int yIn, boolean raw, @Nullable MicListener micListener) {
         super(xIn, yIn, MICROPHONE, null, null, null);
+        this.raw = raw;
         this.micListener = micListener;
         this.client = ClientManager.getClient();
         active = client == null || client.getSoundManager() != null;
@@ -43,9 +46,17 @@ public class MicTestButton extends ToggleImageButton implements ImageButton.Tool
         tooltipSupplier = this;
     }
 
+    public MicTestButton(int xIn, int yIn, boolean raw) {
+        this(xIn, yIn, raw, null);
+    }
+
     @Override
     public void render(PoseStack matrixStack, int x, int y, float partialTicks) {
         super.render(matrixStack, x, y, partialTicks);
+        updateLastRender();
+    }
+
+    public void updateLastRender() {
         if (visible && voiceThread != null) {
             voiceThread.updateLastRender();
         }
@@ -62,6 +73,10 @@ public class MicTestButton extends ToggleImageButton implements ImageButton.Tool
 
     public boolean isHovered() {
         return isHovered;
+    }
+
+    public boolean isMicActive() {
+        return micActive;
     }
 
     @Override
@@ -168,20 +183,27 @@ public class MicTestButton extends ToggleImageButton implements ImageButton.Tool
                 if (System.currentTimeMillis() - lastRender > 500L) {
                     break;
                 }
-                short[] buff = micThread.pollProcessedAudio(true);
+                if (micThread.isClosed()) {
+                    break;
+                }
+                short[] buff = raw ? micThread.pollMic() : micThread.pollProcessedAudio(true);
                 if (buff == null) {
                     continue;
                 }
 
-                micListener.onMicValue(AudioUtils.getHighestAudioLevel(buff));
+                if (micListener != null) {
+                    micListener.onMicValue(AudioUtils.getHighestAudioLevel(buff));
+                }
 
-                if (micThread.shouldTransmitAudio()) {
+                if (raw || micThread.shouldTransmitAudio()) {
                     play(buff);
                 }
             }
             speaker.close();
             setMicLocked(false);
-            micListener.onStop();
+            if (micListener != null) {
+                micListener.onStop();
+            }
             if (usesOwnMicThread) {
                 micThread.close();
             }